backend/requirements.txt
flask==3.0.2
flask-pymongo==2.3.0
flask-cors==4.0.0
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
pymongo==4.6.1
python-dateutil==2.8.2
